solar-panel-detection/panel-detection/requirements.txt
-i https://pypi.org/simple
blinker==1.9.0; python_version >= '3.9'
certifi==2025.1.31; python_version >= '3.6'
charset-normalizer==3.4.1; python_version >= '3.7'
click==8.1.8; python_version >= '3.7'
contourpy==1.3.1; python_version >= '3.10'
cycler==0.12.1; python_version >= '3.8'
filelock==3.18.0; python_version >= '3.9'
flask==3.1.0; python_version >= '3.9'
fonttools==4.56.0; python_version >= '3.8'
fsspec==2025.3.0; python_version >= '3.8'
idna==3.10; python_version >= '3.6'
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.6; python_version >= '3.7'
kiwisolver==1.4.8; python_version >= '3.10'
markupsafe==3.0.2; python_version >= '3.9'
matplotlib==3.10.1; python_version >= '3.10'
mpmath==1.3.0
networkx==3.4.2; python_version >= '3.10'
numpy==2.1.1; python_version >= '3.10'
nvidia-cublas-cu12==12.4.5.8; python_version >= '3'
nvidia-cuda-cupti-cu12==12.4.127; python_version >= '3'
nvidia-cuda-nvrtc-cu12==12.4.127; python_version >= '3'
nvidia-cuda-runtime-cu12==12.4.127; python_version >= '3'
nvidia-cudnn-cu12==9.1.0.70; python_version >= '3'
nvidia-cufft-cu12==11.2.1.3; python_version >= '3'
nvidia-curand-cu12==10.3.5.147; python_version >= '3'
nvidia-cusolver-cu12==11.6.1.9; python_version >= '3'
nvidia-cusparse-cu12==12.3.1.170; python_version >= '3'
nvidia-cusparselt-cu12==0.6.2; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-nccl-cu12==2.21.5; python_version >= '3'
nvidia-nvjitlink-cu12==12.4.127; python_version >= '3'
nvidia-nvtx-cu12==12.4.127; python_version >= '3'
opencv-python==4.11.0.86; python_version >= '3.6'
packaging==24.2; python_version >= '3.8'
pandas==2.2.3; python_version >= '3.9'
pillow==11.1.0; python_version >= '3.9'
psutil==7.0.0; python_version >= '3.6'
py-cpuinfo==9.0.0
pyparsing==3.2.3; python_version >= '3.9'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
pytz==2025.2
pyyaml==6.0.2; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
scipy==1.15.2; python_version >= '3.10'
seaborn==0.13.2; python_version >= '3.8'
setuptools==78.1.0; python_version >= '3.9'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sympy==1.13.1; python_version >= '3.8'
torch==2.6.0; python_full_version >= '3.9.0'
torchvision==0.21.0; python_version >= '3.9'
tqdm==4.67.1; python_version >= '3.7'
triton==3.2.0; platform_system == 'Linux' and platform_machine == 'x86_64'
typing-extensions==4.13.0; python_version >= '3.8'
tzdata==2025.2; python_version >= '2'
ultralytics==8.3.97; python_version >= '3.8'
ultralytics-thop==2.0.14; python_version >= '3.8'
urllib3==2.3.0; python_version >= '3.9'
werkzeug==3.1.3; python_version >= '3.9'
